Cloud Authentication

Firebase.cloudstore_get

Firebase.cloudstore_init

Firebase.cloudstore_sendfile

Firebase.confirm_passwordreset

Firebase.firebase_changeemail

Firebase.firebase_changepassword

Firebase.firebase_confirmemailverify

Firebase.firebase_deleteuser

Firebase.firebase_fetchproviders

Firebase.firebase_getuserinfo

Firebase.firebase_linkoauth

Firebase.firebase_linkuserinfo

Firebase.firebase_passwordreset

Firebase.firebase_signin

Firebase.firebase_signinanon

Firebase.firebase_signinoauth

Firebase.firebase_signup

Firebase.firebase_unlinkprovide

Firebase.firebase_updateprofile

Firebase.firebase_verifyemail

Firebase.firestore_batchget

Firebase.firestore_batchwrite

Firebase.firestore_beginTransaction

Firebase.firestore_commit

Firebase.firestore_createdoc

Firebase.firestore_deletedoc

Firebase.firestore_getcollectionids

Firebase.firestore_getdoclist

Firebase.firestore_partitionquery

Firebase.firestore_patch

Firebase.firestore_runquery

Firebase.get_request

Firebase.readdb_download

Firebase.realdb_delete

Firebase.realdb_get

Firebase.realdb_init

Firebase.realdb_patch

Firebase.realdb_post

Firebase.realdb_put

Firebase.set_webapikey

Firebase.verify_passwordresetcode

Firebase.confirm_passwordreset — Method

Firebase.firebase_changeemail — Function

firebasechangeemail(newemail="",firebaseidtoken= FIREBASEID_TOKEN)

Change email You can change a user's email by issuing an HTTP POST request to the Auth setAccountInfo endpoint.

Example

julia data = firebase_signin("ab669522@gmail.com,"helloworld!!") firebase_changeemail("ab6695221@gmail.com",data["idToken"]) data = firebase_signin("ab6695221@gmail.com,"helloworld!!") firebase_changeemail("ab669522@gmail.com",data["idToken"])`

Firebase.firebase_changepassword — Function

firebasechangepassword(newpassword="", firebaseidtoken=FIREBASEID_TOKEN)

Change password You can change a user's password by issuing an HTTP POST request to the Auth setAccountInfo endpoint.

Example

julia data = firebase_signin("ab669522@gmail.com","helloworld!!") firebase_changepassword("hellopaaji!!",data["idToken"]) data = firebase_signin("ab6695221@gmail.com","hellopaaji!!") firebase_changepassword("helloworld!!",data["idToken"])`

Firebase.firebase_confirmemailverify — Function

TODO: This does work yet

Firebase.firebase_deleteuser — Function

firebasedeleteuser(firebaseidtoken=FIREBASEID_TOKEN)

Delete account You can delete a current user by issuing an HTTP POST request to the Auth deleteAccount endpoint.

Example

data = firebase_signup("ashwani121231@gmail.com","hello!!123")
firebase_deleteuser(data["idToken"])

Firebase.firebase_fetchproviders — Function

firebase_fetchproviders(email="")

Fetch providers for email You can look all providers associated with a specified email by issuing an HTTP POST request to the Auth createAuthUri endpoint.

Example

julia> firebase_fetchproviders("ashwani1@gmail.com")
#= 
Dict{String, Any} with 5 entries:
  "allProviders"  => Any["password"]
  "kind"          => "identitytoolkit#CreateAuthUriResponse"
  "sessionId"     => "k8iBdCSLEeAbhhgKz6E4PC-ltCs"
  "signinMethods" => Any["password"]
  "registered"    => true
=#

Firebase.firebase_getuserinfo — Function

firebasegetuserinfo(firebaseidtoken=FIREBASEID_TOKEN)

Get user data You can get a user's data by issuing an HTTP POST request to the Auth getAccountInfo endpoint.

Example

data = firebase_signin("ab669522@gmail.com","helloworld!!")
firebase_getuserinfo(data["idToken"])

Firebase.firebase_linkoauth — Method

Link with OAuth credential You can link an OAuth credential to a user by issuing an HTTP POST request to the Auth verifyAssertion endpoint.

Firebase.firebase_linkuserinfo — Function

Link with email/password You can link an email/password to a current user by issuing an HTTP POST request to the Auth setAccountInfo endpoint.

Example

data = firebase_signin("ab669522@gmail.com","helloworld!!")
firebase_linkuserinfo("ash@gmail.com", "hello!!123as",data["idToken"])
data = firebase_signin("ash@gmail.com", "hello!!123as")
firebase_linkuserinfo("ab669522@gmail.com","helloworld!!",data["idToken"])

Firebase.firebase_passwordreset — Function

firebase_passwordreset(email="ab669522@gmail.com")

Send password reset email You can send a password reset email by issuing an HTTP POST request to the Auth getOobConfirmationCode endpoint.

Example

firebase_passwordreset(email="ab669522@gmail.com")
# email will recieve the reset password email

Firebase.firebase_signin — Function

firebase_signin(email::String="data", password::String="data", returnSecureToken::Bool= true)

Sign in with email / password You can sign in a user with an email and password by issuing an HTTP POST request to the Auth verifyPassword endpoint.

Example

firebase_signin("ashwani1@gmail.com" , "hello1231!!")
#=
Dict{String, Any} with 8 entries:
  "refreshToken" => "AGEhc0C1F9smilhBPdT5TYClumCwdzGI9q43s5r-MsqW-bJTeiIw1doG7zRF_VMX2a0WELR3zRB…
  "kind"         => "identitytoolkit#VerifyPasswordResponse"
  "localId"      => "zgrI4g3ualVJW7IEiRlPFg5jItA2"
  "displayName"  => ""
  "idToken"      => "eyJhbGciOiJSUzI1NiIsImtpZCI6ImRjNGQwMGJjM2NiZWE4YjU0NTMzMWQxZjFjOTZmZDRlNjd…
  "registered"   => true
  "expiresIn"    => "3600"
  "email"        => "ashwani1@gmail.com"
=#

Firebase.firebase_signinanon — Function

firebase_signinanon(returnSecureToken::Bool= true)

Sign in anonymously You can sign in a user anonymously by issuing an HTTP POST request to the Auth signupNewUser endpoint.

Exampleemail::String="", password::String="",

julia> firebase_signinanon()
#=
Dict{String, Any} with 5 entries:
  "refreshToken" => "AGEhc0DiE1zelLdCW2KG9wMRwZKcZI6EuSef4hsIc6uOZp0I3SYUTf0Hib_XgvtgY1yeqgFNqJ0…
  "kind"         => "identitytoolkit#SignupNewUserResponse"
  "localId"      => "3T8r7DD99zO9qKL1k56boC2iiXF2"
  "expiresIn"    => "3600"
  "idToken"      => "eyJhbGciOiJSUzI1NiIsImtpZCI6ImRjNGQwMGJjM2NiZWE4YjU0NTMzMWQxZjFjOTZmZDRlNjd…
=#

Firebase.firebase_signinoauth — Method

firebase_signinoauth()

Sign in with OAuth credential You can sign in a user with an OAuth credential by issuing an HTTP POST request to the Auth verifyAssertion endpoint.

This doesn't work today

Example

firebase_signinoauth()

Firebase.firebase_signup — Function

firebase_signup(email::String="", password::String="", returnSecureToken::Bool= true)

Sign up with email / password You can create a new email and password user by issuing an HTTP POST request to the Auth signupNewUser endpoint.

Example

set_webapikey("WEB_API_KEY")
firebase_signup("ashwani@gmail.com","hello!!")
#=
Dict{String, Any} with 6 entries:
  "refreshToken" => "AGEhc0AjgXJevwy_rkJ3r4VUHc4EEORngVF-rwEtOuO007AymRqbAzVnn1iAMCZNHqzwnCI1X8fkJ0OPE2fTuGN7ameU6P7SmibwIFs8sOZxUI8C_uklOkk1IG5DFdWXAC45mK_0Hy4tuJXJUvI8Kuy6LS-GERZ5Bc…
  "kind"         => "identitytoolkit#SignupNewUserResponse"
  "localId"      => "GRCBj28Pkhat0mYsJS0BVrlVmJd2"
  "expiresIn"    => "3600"
  "idToken"      => "eyJhbGciOiJSUzI1NiIsImtpZCI6ImRjNGQwMGJjM2NiZWE4YjU0NTMzMWQxZjFjOTZmZDRlNjdjNTFlODkiLCJ0eXAiOiJKV1QifQ.eyJpc3MiOiJodHRwczovL3NlY3VyZXRva2VuLmdvb2dsZS5jb20vZmlyLWp…
  "email"        => "ashwani@gmail.com"
=#

Firebase.firebase_unlinkprovide — Function

Unlink provider You can unlink a provider from a current user by issuing an HTTP POST request to the Auth setAccountInfo endpoint.

Example

set_webapikey()

data = firebase_signup("ashwani12123@gmail.com","hello!!123")
#=
Dict{String, Any} with 6 entries:
  "refreshToken" => "AGEhc0A4Y-wFC0S2Cub-7TY2XNfZMYQf-YKzTdQWl9_SjrRpUW4d3QymQy5P1EN4Me8-SE695sPE_MBXoMoqTzKLy76nBvDGmybGmP4wnWLyrJql-HfoVPRtG1emeJ3E61DnDbUIZOvV-Y3GuzpJ57XZojHBM1aoMp…
  "kind"         => "identitytoolkit#SignupNewUserResponse"
  "localId"      => "neXn8Mg562VFgENRtA2SEC0NEBk1"
  "expiresIn"    => "3600"
  "idToken"      => "eyJhbGciOiJSUzI1NiIsImtpZCI6ImRjNGQwMGJjM2NiZWE4YjU0NTMzMWQxZjFjOTZmZDRlNjdjNTFlODkiLCJ0eXAiOiJKV1QifQ.eyJpc3MiOiJodHRwczovL3NlY3VyZXRva2VuLmdvb2dsZS5jb20vZmlyLWp…
  "email"        => "ashwani12123@gmail.com"
=#
firebase_unlinkprovide(data["idToken"], "google.com")
#=
Dict{String, Any} with 6 entries:
  "kind"             => "identitytoolkit#SetAccountInfoResponse"
  "localId"          => "neXn8Mg562VFgENRtA2SEC0NEBk1"
  "passwordHash"     => "UkVEQUNURUQ="
  "emailVerified"    => false
  "providerUserInfo" => Any[Dict{String, Any}("rawId"=>"ashwani12123@gmail.com", "providerId"=>"password", "federatedId"=>"ashwani12123@gmail.com", "email"=>"ashwani12123@gmail.com")]
  "email"            => "ashwani12123@gmail.com"
=#

Firebase.firebase_updateprofile — Function

Update profile You can update a user's profile (display name / photo URL) by issuing an HTTP POST request to the Auth setAccountInfo endpoint.

TODO: delete attribute

Example

julia data = firebase_signin("ab669522@gmail.com","helloworld!!") firebase_updateprofile("ash","https://i.imgur.com/zgDkgSR.jpeg",data["idToken"]) #= Dict{String, Any} with 8 entries: "kind" => "identitytoolkit#SetAccountInfoResponse" "photoUrl" => "https://i.imgur.com/zgDkgSR.jpeg" "localId" => "GovyExPqFWS70uynnuBzVnYgLAi2" "displayName" => "ash" "passwordHash" => "UkVEQUNURUQ=" "emailVerified" => false "providerUserInfo" => Any[Dict{String, Any}("rawId"=>"ab669522@gmail.com", "photoUrl"=>"https:… "email" => "ab669522@gmail.com" =#

Firebase.firebase_verifyemail — Function

firebaseverifyemail(firebaseidtoken=FIREBASEID_TOKEN)

Send email verification You can send an email verification for the current user by issuing an HTTP POST request to the Auth getOobConfirmationCode endpoint.

Example

data = firebase_signin("ab669522@gmail.com","helloworld!!")
firebase_verifyemail((data["idToken"])

Firebase.set_webapikey — Function

set_webapikey(key ="")

sets WEBAPIKEY

Example

set_webapikey("asjdlajsj1j32lj312")

Firebase.verify_passwordresetcode — Method

verify_passwordresetcode()

Verify password reset code You can verify a password reset code by issuing an HTTP POST request to the Auth resetPassword endpoint.

Example