Cloud Firestore

Firebase.cloudstore_get

Firebase.cloudstore_init

Firebase.cloudstore_sendfile

Firebase.confirm_passwordreset

Firebase.firebase_changeemail

Firebase.firebase_changepassword

Firebase.firebase_confirmemailverify

Firebase.firebase_deleteuser

Firebase.firebase_fetchproviders

Firebase.firebase_getuserinfo

Firebase.firebase_linkoauth

Firebase.firebase_linkuserinfo

Firebase.firebase_passwordreset

Firebase.firebase_signin

Firebase.firebase_signinanon

Firebase.firebase_signinoauth

Firebase.firebase_signup

Firebase.firebase_unlinkprovide

Firebase.firebase_updateprofile

Firebase.firebase_verifyemail

Firebase.firestore_batchget

Firebase.firestore_batchwrite

Firebase.firestore_beginTransaction

Firebase.firestore_commit

Firebase.firestore_createdoc

Firebase.firestore_deletedoc

Firebase.firestore_getcollectionids

Firebase.firestore_getdoclist

Firebase.firestore_partitionquery

Firebase.firestore_patch

Firebase.firestore_runquery

Firebase.get_request

Firebase.readdb_download

Firebase.realdb_delete

Firebase.realdb_get

Firebase.realdb_init

Firebase.realdb_patch

Firebase.realdb_post

Firebase.realdb_put

Firebase.set_webapikey

Firebase.verify_passwordresetcode

Firebase.firestore_batchget — Function

firestore_batchget(batch=body)

Gets multiple documents Documents returned by this method are not guaranteed to be returned in the same order that they were requested.

Example

init("[FIREBASE_ADMIN_SDK].json")
res = firestore_batchget()

Firebase.firestore_batchwrite — Function

firestore_batchwrite(batch=body)

Applies a batch of write operations. The documents.batchWrite method does not apply the write operations atomically and can apply them out of order. Method does not allow more than one write per document. Each write succeeds or fails independently. See the BatchWriteResponse for the success status of each write.

If you require an atomically applied set of writes, use documents.commit instead.

Example

init("[FIREBASE_ADMIN_SDK].json")
res = firestore_batchwrite()

Firebase.firestore_beginTransaction — Method

firestore_beginTransaction()

Starts a new transaction. "options" The options for the transaction. Defaults to a read-write transaction.

Example

init("[FIREBASE_ADMIN_SDK].json")
firestore_beginTransaction()

Firebase.firestore_commit — Function

firestore_commit(transactionid="",body =body)

Commits a transaction, while optionally updating documents.

Example

init("[FIREBASE_ADMIN_SDK].json")
data = firestore_beginTransaction()
firestore_commit(data["transaction"])

Firebase.firestore_createdoc — Function

firestore_createdoc(url, body)

Example

init("[FIREBASE_ADMIN_SDK].json")
firestore_createdoc("/firebase_test/firebase_post/two") 
firestore_createdoc("/firebase_test/firebase_post/three";query ="hello") # with document id name in query

Firebase.firestore_deletedoc — Method

firestore_deletedoc(url)

Deletes a document

Example

init("[FIREBASE_ADMIN_SDK].json")
firestore_createdoc("/firebase_test/firebase_post/three";query ="hello") # creates the document and then we delete it
firestore_deletedoc("/firebase_test/firebase_post/three/hello")

Firebase.firestore_getcollectionids — Method

firestore_getcollectionids(url;query="")

Lists all the collection IDs underneath a document.

Example

init("[FIREBASE_ADMIN_SDK].json")
firestore_getcollectionids("/firebase_test/firebase_post")

Firebase.firestore_getdoclist — Method

firestore_getdoclist(url)

Lists documents

Example

init("[FIREBASE_ADMIN_SDK].json")
firestore_getdoclist("/firebase_test")

Firebase.firestore_partitionquery — Function

firestore_partitionquery(url)

Partitions a query by returning partition cursors that can be used to run the query in parallel. The returned partition cursors are split points that can be used by documents.runQuery as starting/end points for the query results.

Firebase.firestore_patch — Function

firestore_patch()

Updates or inserts a document.

Example

init("[FIREBASE_ADMIN_SDK].json")
firestore_patch("/firebase_test/firebase_patch")

Firebase.firestore_runquery — Function

firestore_runquery()

Runs a query.

Example

init("[FIREBASE_ADMIN_SDK].json")
firestore_runquery("/firebase_test/firebase_post")

Firebase.get_request — Method

get_request(url::String)

Get request which is general in nature and gets any document or collection

Example

init("[FIREBASE_ADMIN_SDK].json")
res = get_request("/firebase_test/firebase_get") # document get
res = get_request("/firebase_test/firebase_get/firebase_get_collection") # collection fetch