Realtime Database

Firebase.cloudstore_get

Firebase.cloudstore_init

Firebase.cloudstore_sendfile

Firebase.confirm_passwordreset

Firebase.firebase_changeemail

Firebase.firebase_changepassword

Firebase.firebase_confirmemailverify

Firebase.firebase_deleteuser

Firebase.firebase_fetchproviders

Firebase.firebase_getuserinfo

Firebase.firebase_linkoauth

Firebase.firebase_linkuserinfo

Firebase.firebase_passwordreset

Firebase.firebase_signin

Firebase.firebase_signinanon

Firebase.firebase_signinoauth

Firebase.firebase_signup

Firebase.firebase_unlinkprovide

Firebase.firebase_updateprofile

Firebase.firebase_verifyemail

Firebase.firestore_batchget

Firebase.firestore_batchwrite

Firebase.firestore_beginTransaction

Firebase.firestore_commit

Firebase.firestore_createdoc

Firebase.firestore_deletedoc

Firebase.firestore_getcollectionids

Firebase.firestore_getdoclist

Firebase.firestore_partitionquery

Firebase.firestore_patch

Firebase.firestore_runquery

Firebase.get_request

Firebase.readdb_download

Firebase.realdb_delete

Firebase.realdb_get

Firebase.realdb_init

Firebase.realdb_patch

Firebase.realdb_post

Firebase.realdb_put

Firebase.set_webapikey

Firebase.verify_passwordresetcode

Realtime Database

Initializing the realdb for the base url, reduces the effort later

realdbinit("https://fir-jl-457eb-default-rtdb.asia-southeast1.firebasedatabase.app")
Base Url set:"https://fir-jl-457eb-default-rtdb.asia-southeast1.firebasedatabase.app"

GET - Reading Data

Data from your Realtime Database can be read by issuing an HTTP GET request to an endpoint. The following example demonstrates how you might retrieve a user's name that you had previously stored in Realtime Database.

julia> realdb_get("/firebase_test")
ulia> realdb_get("/firebase_test")
FINAL URL: "https://fir-jl-457eb-default-rtdb.asia-southeast1.firebasedatabase.app/firebase_test.json"
GET successful
Dict{String, Any} with 3 entries:
  "firebase_get"  => Dict{String, Any}("firebase_get_3"=>3, "firebase_get_2"=>2, "firebase_get_1"=>1)
  "firebase_new"  => Dict{String, Any}("-McEjFxupzSbZwojItzL"=>Dict{String, Any}("name"=>"real_db_test"), "-McEjCHubimhfipKm2yo"=>Dict{String, Any}("name"=>"real_db_test"), "-McEjNy8-8VjqSH3P8dO"=>Dict{String, Any}…
  "firebase_post" => Dict{String, Any}("firebase_post_1"=>1, "firebase_post_3"=>3, "firebase_post_2"=>2, "input"=>Dict{String, Any}("-McEQ7vGcvji7Myab182"=>Dict{String, Any}("stockAmount"=>350, "price"=>"7.5 TL", "…

POST - Pushing Data

To accomplish the equivalent of the JavaScript push() method (see Lists of Data), you can issue a POST request. A successful request is indicated by a 200 OK HTTP status code. The response contains the child name of the new data specified in the POST request.

julia> realdb_post("/firebase_test/firebase_new")
FINAL URL: "https://fir-jl-457eb-default-rtdb.asia-southeast1.firebasedatabase.app/firebase_test/firebase_new.json"
POST successful
Dict{String, Any} with 1 entry:
  "name" => "-McEkk8qvdT4UHSUQALx"

julia> realdb_post("/firebase_test/firebase_new/","""{"hello":1}""")
FINAL URL: "https://fir-jl-457eb-default-rtdb.asia-southeast1.firebasedatabase.app/firebase_test/firebase_new.json"
POST successful
Dict{String, Any} with 1 entry:
  "name" => "-McElRx3ZBP-l7pCF0LN"

DELETE - Removing Data

You can delete data with a DELETE request A successful DELETE request is indicated by a 200 OK HTTP status code with a response containing JSON null.

julia> realdb_delete("/firebase_test/firebase_new")
https://fir-jl-457eb-default-rtdb.asia-southeast1.firebasedatabase.app/firebase_test/firebase_new.json
DELETE successful

PUT - Writing Data

You can write data with a PUT request. A successful request is indicated by a 200 OK HTTP status code. The response contains the data specified in the PUT request.

julia> realdb_put("/firebase_test/firebase_new/","""{"hello":1}""")
FINAL URL: "https://fir-jl-457eb-default-rtdb.asia-southeast1.firebasedatabase.app/firebase_test/firebase_new.json"
PUT successful
Dict{String, Any} with 1 entry:
  "name" => "-McElRx3ZBP-l7pCF0LN"

Firebase.readdb_download — Function

readdb_download(url, filename = "test"; query = Dict())

Download request

Example

Firebase.realdb_delete — Function

realdbdelete(url, body = Dict("name"=> "realdb_test"); query = Dict())

DELETE request to an endpoint

Example

realdb_init("https://[PROJECT_ID].asia-southeast1.firebasedatabase.app")
realdb_delete("/users/jack/name/last")

Firebase.realdb_get — Method

realdb_get(url; query = Dict())

GET request to an endpoint.

Example

realdb_init("https://[PROJECT_ID].asia-southeast1.firebasedatabase.app")
realdb_get("/users/jack/name")

Firebase.realdb_init — Method

realdbinit(baseurl; query = Dict())

Initialize the realtimedb with baseurl to make things easier

Example

realdb_init("https://[PROJECT_ID].asia-southeast1.firebasedatabase.app")

Firebase.realdb_patch — Function

realdbpatch(url, body = Dict("name"=> "realdb_test"); query = Dict())

PATCH request to an endpoint.

Example

realdb_init("https://[PROJECT_ID].asia-southeast1.firebasedatabase.app")
body =Dict("last"=>"Jones")
realdb_patch("/users/jack/name/",body)

Firebase.realdb_post — Function

realdbpost(url, body = "{"name": "realdb_test"}"; query = Dict())

POST request to an endpoint.

Example

realdb_init("https://[PROJECT_ID].asia-southeast1.firebasedatabase.app")
# realdb_post("/message_list")
body =Dict("user_id" => "jack", "text" => "Ahoy!")
realdb_post("/message_list",body)

Notes:

According to the REST API documentation for Realtime Database, a POST is the equivalent of a "push" operation when using the client SDKs. This push operation always involves adding data under a random ID. There is no avoiding that for a push.

If you know the name of the node where you want to add data, you should use a PUT instead. This is the equivalent of using "set" operation with the client SDKs.

Firebase.realdb_put — Function

realdbput(url, body = Dict("name"=> "realdb_test"); query = Dict())

PUT request to a endpoint.

Example

realdb_init("https://[PROJECT_ID].asia-southeast1.firebasedatabase.app")
body = Dict("first"=>"Ash", "last"=>"Sparrow")
realdb_put("/users/jack/name",body)