Utility Functions

Firebase.cloudstore_get

Firebase.cloudstore_init

Firebase.cloudstore_sendfile

Firebase.confirm_passwordreset

Firebase.firebase_changeemail

Firebase.firebase_changepassword

Firebase.firebase_confirmemailverify

Firebase.firebase_deleteuser

Firebase.firebase_fetchproviders

Firebase.firebase_getuserinfo

Firebase.firebase_linkoauth

Firebase.firebase_linkuserinfo

Firebase.firebase_passwordreset

Firebase.firebase_signin

Firebase.firebase_signinanon

Firebase.firebase_signinoauth

Firebase.firebase_signup

Firebase.firebase_unlinkprovide

Firebase.firebase_updateprofile

Firebase.firebase_verifyemail

Firebase.firestore_batchget

Firebase.firestore_batchwrite

Firebase.firestore_beginTransaction

Firebase.firestore_commit

Firebase.firestore_createdoc

Firebase.firestore_deletedoc

Firebase.firestore_getcollectionids

Firebase.firestore_getdoclist

Firebase.firestore_partitionquery

Firebase.firestore_patch

Firebase.firestore_runquery

Firebase.get_request

Firebase.init

Firebase.initparsedjson

Firebase.readdb_download

Firebase.realdb_delete

Firebase.realdb_get

Firebase.realdb_init

Firebase.realdb_patch

Firebase.realdb_post

Firebase.realdb_put

Firebase.set_webapikey

Firebase.verify_passwordresetcode

Firebase.init — Method

init( julianparseddict::Dict)

Initialize if the sdk key is in Julian dict format and the file is saved as adminsdkkey.json

Example

```md julia> r = HTTP.get(BASE_URL, header); # returns a response with header and body HTTP.Messages.Response

julia> response = JSON.parse(String(r.body)) Dict{String, Any} with 10 entries: "clientemail" => "****" "clientid" => "****" "authuri" => "https://accounts.google.com/o/oauth2/auth" "authproviderx509certurl" => "https://www.googleapis.com/oauth2/v1/certs" "tokenuri" => "https://oauth2.googleapis.com/token" "clientx509certurl" => "https://www.googleapis.com/robot/v1/metadata/x509/firebase-adminsdk-jk50g%40volka-bot.iam.gserviceaccount.com" "projectid" => "****" "type" => "serviceaccount" "privatekey" => "****" "privatekeyid" => "****"

julia> init(response) Project Admin SDK key is adminsdkkey.json now!! projectid: **** clientemail: **** ````

Firebase.initparsedjson — Method

initparsedjson( parsedjson)

initialize if the sdk key is in parsed JSON format and the file is saved as adminsdkkey.json

Example

```md julia> r = HTTP.get(BASE_URL, header); # returns a response with header and body HTTP.Messages.Response

julia> response = JSON.json(JSON.parse(String(r.body))) "{"clientemail":"****","clientid":"****","authuri":"https://accounts.google.com/o/oauth2/auth","authproviderx509certurl":"https://www.googleapis.com/oauth2/v1/certs","tokenuri":"https://oauth2.googleapis.com/token","clientx509certurl":"https://www.googleapis.com/robot/v1/metadata/x509/firebase-adminsdk-jk50g%40volka-bot.iam.gserviceaccount.com","projectid":"volka-bot","type":"serviceaccount","privatekey":"****","privatekeyid":"****"}"

julia> initparsedjson( response) Project Admin SDK key is adminsdkkey.json now!! projectid: **** clientemail: **** ````